--- modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-renderer/src/main/resources/META-INF/resources/js/store/actions/handleFieldFocused.es.js
@@ -1,31 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-import {PagesVisitor} from '../../util/visitors.es';
-
-export default (pages, properties) => {
-	const {fieldInstance} = properties;
-	const pageVisitor = new PagesVisitor(pages);
-
-	return Promise.resolve(
-		pageVisitor.mapFields((field) => {
-			const focused = field.name === fieldInstance.name;
-
-			return {
-				...field,
-				focused,
-			};
-		})
-	);
-};
